Function to convert the grid into an array
Function to turn the grid area into a 2d numpy array and return a basic direction for the snakehead to reach the food, from here we just need to find out how to simulate a pygame event in the gameloop funciton. I also added a function to make sure food is not in the way of obstacles, there is. abug with obstacles that the food could spanw within them making it impossible to reach the food.